+++ core-ng/src/main/java/core/framework/internal/db/RepositoryImpl.java
@@ -89,10 +89,10 @@ public final class RepositoryImpl<T> implements Repository<T> {
     }
 
     @Override
-    public void update(T entity) {
+    public boolean update(T entity) {
         int updatedRows = update(entity, false, null, null);
-        if (updatedRows != 1)
-            logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
+        if (updatedRows != 1) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
+        return updatedRows == 1;
     }
 
     private int update(T entity, boolean partial, String where, Object[] params) {
@@ -112,31 +112,29 @@ public final class RepositoryImpl<T> implements Repository<T> {
     }
 
     @Override
-    public void partialUpdate(T entity) {
+    public boolean partialUpdate(T entity) {
         int updatedRows = update(entity, true, null, null);
-        if (updatedRows != 1)
-            logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
+        if (updatedRows != 1) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 1, rows={}", updatedRows);
+        return updatedRows == 1;
     }
 
     @Override
     public boolean partialUpdate(T entity, String where, Object... params) {
         int updatedRows = update(entity, true, where, params);
-        if (updatedRows > 1)
-            logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "updated rows is not 0 or 1, rows={}", updatedRows);
         return updatedRows == 1;
     }
 
     @Override
-    public void delete(Object... primaryKeys) {
+    public boolean delete(Object... primaryKeys) {
         var watch = new StopWatch();
         if (primaryKeys.length != selectQuery.primaryKeyColumns) {
-            throw new Error(Strings.format("the length of primary keys does not match columns, primaryKeys={}, columns={}", selectQuery.primaryKeyColumns, primaryKeys.length));
+            throw new Error(Strings.format("the length of primary keys must match columns, primaryKeys={}, columns={}", primaryKeys.length, selectQuery.primaryKeyColumns));
         }
         int deletedRows = 0;
         try {
             deletedRows = database.operation.update(deleteSQL, primaryKeys);
-            if (deletedRows != 1)
-                logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", deletedRows);
+            if (deletedRows != 1) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows is not 1, rows={}", deletedRows);
+            return deletedRows == 1;
         } finally {
             long elapsed = watch.elapsed();
             int operations = ActionLogContext.track("db", elapsed, 0, deletedRows);
@@ -170,6 +168,7 @@ public final class RepositoryImpl<T> implements Repository<T> {
     public boolean[] batchInsertIgnore(List<T> entities) {
         var watch = new StopWatch();
         if (entities.isEmpty()) throw new Error("entities must not be empty");
+        if (insertQuery.generatedColumn != null) throw new Error("entity must not have auto increment primary key, entityClass=" + entityClass.getCanonicalName());
         String sql = insertQuery.insertIgnoreSQL();
         List<Object[]> params = new ArrayList<>(entities.size());
         for (T entity : entities) {
@@ -180,7 +179,7 @@ public final class RepositoryImpl<T> implements Repository<T> {
         try {
             int[] results = database.operation.batchUpdate(sql, params);
             boolean[] insertedResults = new boolean[results.length];
-            insertedRows = insertedResults(results, insertedResults);
+            insertedRows = updatedResults(results, insertedResults);
             return insertedResults;
         } finally {
             long elapsed = watch.elapsed();
@@ -190,24 +189,8 @@ public final class RepositoryImpl<T> implements Repository<T> {
         }
     }
 
-    // use procedurally way to produce results to insertedRows and insertedResults in one method, to balance both performance and ease of unit test
-    // not recommended applying in application level code
-    int insertedResults(int[] results, boolean[] insertedResults) {
-        int insertedRows = 0;
-        for (int i = 0; i < results.length; i++) {
-            int result = results[i];
-            // with insertIgnore, mysql returns -2 if insert succeeds, for batch only
-            // refer to com.mysql.cj.jdbc.ClientPreparedStatement.executeBatchedInserts
-            if (result == Statement.SUCCESS_NO_INFO || result > 0) {
-                insertedResults[i] = true;
-                insertedRows++;
-            }
-        }
-        return insertedRows;
-    }
-
     @Override
-    public void batchDelete(List<?> primaryKeys) {
+    public boolean[] batchDelete(List<?> primaryKeys) {
         var watch = new StopWatch();
         if (primaryKeys.isEmpty()) throw new Error("primaryKeys must not be empty");
         List<Object[]> params = new ArrayList<>(primaryKeys.size());
@@ -221,14 +204,31 @@ public final class RepositoryImpl<T> implements Repository<T> {
         int deletedRows = 0;
         try {
             int[] results = database.operation.batchUpdate(deleteSQL, params);
-            deletedRows = Arrays.stream(results).sum();
-            if (deletedRows != primaryKeys.size())
-                logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows does not match size of primary keys, rows={}", Arrays.toString(results));
+            boolean[] deletedResults = new boolean[results.length];
+            deletedRows = updatedResults(results, deletedResults);
+            if (deletedRows != primaryKeys.size()) logger.warn(errorCode("UNEXPECTED_UPDATE_RESULT"), "deleted rows do not match size of primary keys, rows={}", Arrays.toString(results));
+            return deletedResults;
         } finally {
             long elapsed = watch.elapsed();
             int operations = ActionLogContext.track("db", elapsed, 0, deletedRows);
             logger.debug("batchDelete, sql={}, params={}, size={}, elapsed={}", deleteSQL, new SQLBatchParams(database.operation.enumMapper, params), primaryKeys.size(), elapsed);
             database.checkOperation(elapsed, operations);
         }
+    }
+
+    // use procedurally way to produce results to insertedRows and insertedResults in one method, to balance both performance and ease of unit test
+    // not recommended applying in application level code
+    int updatedResults(int[] results, boolean[] updatedResults) {
+        int updatedRows = 0;
+        for (int i = 0; i < results.length; i++) {
+            int result = results[i];
+            // with insertIgnore, mysql returns -2 if insert succeeds, for batch only
+            // refer to com.mysql.cj.jdbc.ClientPreparedStatement.executeBatchedInserts
+            if (result == Statement.SUCCESS_NO_INFO || result > 0) {
+                updatedResults[i] = true;
+                updatedRows++;
+            }
+        }
+        return updatedRows;
     }
 }