+++ core-ng/src/test/java/core/framework/internal/db/RepositoryImplTest.java
@@ -19,14 +19,14 @@ class RepositoryImplTest {
     }
 
     @Test
-    void insertedResults() {
+    void updatedResults() {
         boolean[] results = new boolean[3];
-        assertThat(repository.insertedResults(new int[]{Statement.SUCCESS_NO_INFO, 0, Statement.SUCCESS_NO_INFO}, results))
+        assertThat(repository.updatedResults(new int[]{Statement.SUCCESS_NO_INFO, 0, Statement.SUCCESS_NO_INFO}, results))
             .isEqualTo(2);
         assertThat(results).contains(true, false, true);
 
         results = new boolean[2];
-        assertThat(repository.insertedResults(new int[]{1, 1}, results)).isEqualTo(2);
+        assertThat(repository.updatedResults(new int[]{1, 1}, results)).isEqualTo(2);
         assertThat(results).contains(true, true);
     }
 }